rag/requirements.txt
fastapi
uvicorn
pymupdf
pillow
pytesseract
numpy
hnswlib
sentence-transformers
ollama
pydantic